pi/requirements.txt
pyserial
pandas
picamera
RPi.GPIO